add non padded accuracy

+++ distill/models/lstm_seq2seq.py
@@ -129,7 +129,7 @@ class LSTMSeq2Seq(object):
       output_mask = tf.cast(tf.sequence_mask(lstm_decoder_output_dic['outputs_lengths'], tf.shape(outputs)[1]), dtype=tf.int64)
 
       logits = self.output_embedding_layer.linear(outputs)
-      predictions = tf.cast(tf.argmax(logits, axis=-1) * output_mask, dtype=tf.int32)
+      predictions = tf.cast(tf.argmax(logits, axis=-1) * output_mask, dtype=tf.int64)
 
     return {'logits': logits,
             'output_mask': output_mask,
